fix: unpad amounts from chainlink in payment-detection
