+++ packages/payment-detection/package.json
@@ -43,7 +43,7 @@
     "@requestnetwork/currency": "0.8.0",
     "@requestnetwork/smart-contracts": "0.28.0",
     "@requestnetwork/types": "0.35.0",
-    "@requestnetwork/utils": "0.35.0",
+    "@requestnetwork/utils": "^0.35.0",
     "axios": "0.21.1",
     "ethers": "5.1.4",
     "graphql": "15.5.0",
