+++ packages/payment-detection/src/index.ts
@@ -7,6 +7,7 @@ import * as Erc20PaymentNetwork from './erc20';
 import * as EthPaymentNetwork from './eth';
 import { initPaymentDetectionApiKeys, setProviderFactory, getDefaultProvider } from './provider';
 import { getTheGraphClient, networkSupportsTheGraph } from './thegraph';
+import { padAmountForChainlink, unpadAmountFromChainlink } from './utils';
 export type { TheGraphClient } from './thegraph';
 
 export {
@@ -21,4 +22,6 @@ export {
   getDefaultProvider,
   getTheGraphClient,
   networkSupportsTheGraph,
+  padAmountForChainlink,
+  unpadAmountFromChainlink,
 };
